ajout de couleur de order detail fait
