+++ app/Models/Enums/OrderdetailStatus.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Models\Enums;
+
+use Illuminate\Support\Arr;
+
+enum OrderdetailStatus: string implements AdvancedEnumInterface
+{
+    use AdvancedEnum;
+
+    case DISPONIBLE = 'disponible';
+    case INDISPONIBLE = 'indisponible';
+
+    public function label(): string
+    {
+        return __('orderdetail.status.'.$this->value);
+    }
+
+    public static function default(): string
+    {
+        return self::DISPONIBLE->value;
+    }
+
+    public function color(): string
+    {
+        return Arr::get($this->colors(), $this->value);
+    }
+
+    public function colors(): array
+    {
+        return [
+            self::DISPONIBLE->value => '0xFF28A745',
+            self::INDISPONIBLE->value => '0xFFDC3545',
+        ];
+    }
+}
