+++ lang/fr/orderdetail.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'status.disponible' => 'Disponible',
+    'status.indisponible' => 'Indisponible',
+];
