+++ routes/Api/dataRoutes.php
@@ -1,6 +1,6 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PharmacyController;
+use Illuminate\Support\Facades\Route;
 
 Route::get('/filterProduct', [PharmacyController::class, 'getPharFliter']);
